Move transitions into tokens
