+++ libraries/yas-style/src/tokens.ts
@@ -172,3 +172,12 @@ export const easings = {
   legacyDecelerate: "cubic-bezier(0, 0, 0.2, 1)",
   linear: "cubic-bezier(0, 0, 1, 1)",
 };
+
+export const transitions = {
+  emphasizedBeginAndEndOnScreen: `${durations.long2} ${easings.emphasized}`,
+  emphasizedEnter: `${durations.medium4} ${easings.emphasizedDecelerate}`,
+  emphasizedExit: `${durations.short4} ${easings.emphasizedAccelerate}`,
+  standardBeginAndEndOnScreen: `${durations.medium2} ${easings.standard}`,
+  standardEnter: `${durations.medium1} ${easings.standardDecelerate}`,
+  standardExit: `${durations.short4} ${easings.standardAccelerate}`,
+};
